Initial commit
Hypcast is designed to allow for easy remote viewing of signals from a
TV tuner card. I've figured out how to stream live TV from my Linux
desktop (with a Hauppauge tuner card) to my laptop using azap, ffmpeg,
and ssh. However, I'm pretty sure I can do better than this, and I want
an excuse to get familiar with Node (preferably with something new - I
could just redevelop Peribot, but that wouldn't be terribly fun).

Instead of the craziness of dealing with azap and stuff directly,
Hypcast will provide a web interface where a channel can be selected for
encoding as an HLS (HTTP Live Streaming) stream. The video size and
stream quality can be set as well. Ideally, multiple streams would be
provided and the client would choose between them based on network
conditions (which is kind of the point of HLS), but my computer can't
scale and encode 4 video streams simultaneously in real time (and the
encoding DOES need to happen in real time to have a live broadcast!).