Create initial HlsTunerStreamer
Create an initial version of HlsTunerStreamer, which is designed to
manage the creation of HLS streams from a Tuner (such as AzapTuner).
This entire module is based on a state machine that is designed to
ensure that everything gets set up and torn down correctly. This will
probably matter more once the FFmpeg stuff is added.