Remove npm-shrinkwrap
Remove the NPM shrinkwrap file, as it did not really account for all
dependencies and thus did not really serve its desired purpose.

That is to say... a fresh npm install && npm shrinkwrap led to hundreds
of lines being added to this file for dependencies that didn't exist
before. It seems that I was never properly managing it, so I don't see
much good in keeping it around and only having half of my stuff
shrinkwrapped.