Remove jQuery for good
Remove jQuery entirely by adding in a nice HTTP request library that can
be used instead to retrieve channels and profiles. This drops the size
of the non-minified client build by about 0.4 MB. Granted, the client
build is still over 2 MB in size so you'd never normally want to use it
anyways. But I guess it's still nice.