Ensure tuneData is set by default
Fix a bug that occurred when a stream was started without modifying the
channel or profile. These values would be blank by default, which would
cause FFmpeg to crash when it couldn't understand the filters on the
command line (since the video height would be undefined).

I'm accomplishing this by setting up the tuneData to have default values
after channels and profiles load. I'm trying to avoid a race condition
where the server (if tuned) and client overwrite each other. It feels a
bit weird, but I'm also trying to keep the UI component from owning the
tuneData in any way (since we need to be sure that the server can always
override it if needed).