Re-implement video player transitions in React UI
Re-implement the transition of the video player that occurs when it is
shown/hidden. This was part of the original Hypcast UI, but was not
added back in after the React rewrite until now.

The new transition is slightly different, as the player now grows in
height and width instead of just height. Also, the video continues to
play until the transition is complete. However, because the transition
is now implemented entirely in CSS, it looks a lot smoother (at least
subjectively).

This is another really nice use of CSS modules in the Hypcast UI. The
classes used by ReactCSSTransitionGroup can be implemented in CSS and
passed in via JavaScript just by requiring the LESS file and getting an
object with the necessary class names. I think that really helps keep
this clean and organized, vs. setting up class names in CSS and having
to manually keep them in sync with what the React component gets.