Grow/shrink player from center instead of left
Modify the video player transitions so that the player appears from and
disappears into the top center of the surrounding container, rather than
the top left. This is accomplished by setting the player's position to
relative for the duration of the transition and animating the left
property between 0 and 50%.

As part of this, the animation classes now use LESS mixins to define the
endpoints of each transition. This means that the transition endpoint
styles can be reused across the class definitions required by
ReactCSSTransitionGroup. Note that CSS module composition does not work,
because only one animation class can be applied to the element at a
time.