Switch hls_wrap to hls_flags delete_segments
Remove a deprecated FFmpeg option in favor of one that is apparently
better supported. Note that a side effect of this change is older stream
files staying on disk longer and lengthening the amount of the stream
that remains available (or, at least that seems to be the effect from my
informal testing).