Use async/await for full buffering phase
Remove the unnecessary state machine actions in the "buffering" phase of
the streamer, and use async/await for temp directory creation and
playlist file touching. This allows that code to be consolidated inside
of a single try/catch in the _onEnter function, to try to make it a bit
cleaner.