Bump to ESLint 4
Bump up to ESLint 4.1.0. Reconfigure it to continue using legacy
indentation checking, since it refuses to match the indentation style I
want for JSX params (though, maybe my style is dumb and I should just
give up on it). Also, enforce a space between "async" and parens in
arrow functions. That's not dumb.