Upgrade to react-transition-group 2.x
This upgrade gives us a much cleaner way to transition the video state,
since we can use a "CSSTransition" component directly without wrapping
things in a group and creating a fake <span> element.

Note that my initial iterations of this change focused on keeping the
TransitionGroup logic around, based on a fairly straight application of
the migration guide in react-transition-group's README. However, this
caused significant frustration, as I found that the video would stop
playing while it was transitioning out. Eventually I discovered that the
<video> element was being destroyed and recreated during the exit
transition. After running through the bug tracker and R-ing TFM, I
realized that the whole "group" thing wasn't even necessary anymore.

Note that we do explicitly have the transition component (un)mount the
HlsVideoPlayer child as it transitions in and out, since this is how
that component manages resources from hls.js.