Absorb SIGKILL from ffmpeg prior to re-tuning
As part of the recent error handling changes, the "tuning" state gained
an ffmpegError handler that I thought was useless. However, it turns out
that killing ffmpeg causes it to emit an error, which when retuning
invokes the error action on the "tuning" state. I was expecting Machina
to invoke this on the "debuffering" state instead (since that one
doesn't have an ffmpeg error handler).

Anyways... the error was silently lost before and is explicitly dropped
now, so retuning works again.