Vendor deb-multimedia-keyring
It's kind of stupid that I only install signed deb-multimedia
packages... but also install the signing key from a potentially
untrusted repository and ignore signature errors.

The key is now vendored into my repository, so if something *does*
change (in a good or bad way) it'll be obvious. I accessed the
deb-multimedia.org homepage over HTTPS and the SHA256 I saw matched that
of the file in this repository, so hopefully we should be good to go in
that sense.

Also the MAINTAINER in the Dockerfile now includes my email.