Go back to using specific Node versions
Turns out Docker Hub has some weird stuff with publishing, where when a
tag is brand-new some of the architectures it references might not exist
yet, and you get weird errors as a result.

I just switched from aufs to overlay2 on my desktop, so I needed to do
this to get Hypcast running again in the meantime. May as well keep it.