Switch server compilation to TypeScript
This completely removes Babel from Hypcast, and replaces its
functionality with that of TypeScript. We're using tsconfig.json to
represent server settings and the Webpack config to represent client
settings, just like for Babel.

The only code change was to redo imports and match TypeScript's
expectations (`import * as X` rather than `import X` for CommonJS
modules).