Fix server build issue due to shared modules
Becase the `models/` directory was implicitly added to the TypeScript
compilation, the output went from this:

  dist/
    server/
      index.js
      (etc)

to this:

  dist/
    server/
      server/
        index.js
        (etc)
      models/
        TuneData.js

When a fresh Docker image was built, it could no longer find the server
code. This change should fix the issue by having both directories output
into `dist/`.

(As a side note, I'm not sure why tsc emits a TuneData.js at all. It's
just a set of interface declarations.)