Optimize the Docker build for some reason
Docker *has* multi-stage builds so obviously I'm required to take
advantage of them.

More seriously, I didn't realize until now that a stage can appear as
the base in a FROM line, not just as an option to COPY. That helps keep
me from needing to re-install dvb-apps and stuff every time I change a
tiny bit of code and need to rebuild a production-ready image.