Switch back to major-version Docker images for Node
In fee3fcfa, I switched from the "x-slim" tag format for Node images to
the "x.y.z-slim" format. This was done as a workaround for an issue in
Docker Hub where a new "x-slim" tag was published while some of its
architectures were missing, leading to failures on pull. However, now
that I have automatic builds running on Docker Hub with a link to the
"node" repository, Docker Hub might be able to handle this on its own.
Will try it out and see what happens, since this could be a great way to
keep the image up to date with security remediations and whatnot.