Update note about fluent-ffmpeg types
The thing with console logging was fixed in DefinitelyTyped, and never
wrong in the documentation in the first place (at least in the README; I
don't know if I looked at something else or was completely mistaken).

However, there are still a couple other issues preventing my immediate
adoption of these types.