Optimize Docker build for BuildKit
Now that BuildKit is available by default in Docker v18.09+ (assuming
DOCKER_BUILDKIT=1 is set in the client environment), there is a
possibility of massively speeding up multi-stage builds through
parallelization.

This is about as heavily as I can possibly optimize the Dockerfile for
parallelizability and cacheability. In rudimentary testing (docker
system prune followed by yarn run docker:build), the overall build time
was cut roughly in half by this change. The primary driver of time on a
fresh build is the installation of ffmpeg, dvb-apps, etc., which can now
be done at the same time as the JavaScript client and server builds.

Splitting of the client and server builds into different stages is
probably not a _strictly_ necessary step, though it may enable improved
caching. I'm not sure that the time savings are actually significant
enough in this case, and might bring these back into one stage in the
future.