Display 720p video and tweak encoder settings
With this setup we're right around 100% CPU usage while actively
encoding video, with about 40% going to libvpx and most of the rest
going to other places in the pipeline.

Maybe it's time to consider a CPU upgrade if I really want to stream
1080p video? Alternatively I wonder if H.264 or VP9 would fare better.
x264 works pretty well doing live encodes for Hypcast.