Enable dynamic management of GStreamer pipelines
With this change, we can now create individual GStreamer pipelines to
tune to specific channels and destroy them when we're done. This will
enable us to tune channels from the client end without restarting the
server.

It might be better to directly set the frequency and program-id on the
dvbsrc and demuxer elements in the pipeline in order to change channels,
but the current technique seems to reduce the amount of interaction we
need to do with GStreamer (i.e. between Go and C). That said, it might
be worthwhile to modify the API so it *looks* like that's how it works.

The sink callback is facilitated by giving every pipeline a global
unsigned int ID and keeping a map of every pipeline in existence. Will
add more details in a later commit.