Add new, more explicit channel selector
This is a bit closer to the eventual pattern we want for channel
selection, which is to have a list of clickable items where the click
action on a specific channel triggers a change.

It also opens the door to have the channel list interact less with the
control logic, by pulling its own channels.