Initial version of watch.Value
This is a mechanism to create a synchronized value that can be watched
by multiple consumers, guaranteeing that no matter what they do or how
slow they are they will eventually see the latest state of the value.

The tuner code currently includes a bespoke version of this concept,
with significantly less testing and/or reusability. The goal of this
implementation is to replace that.

This implementation has been heavily tested with the race detector, as
well as the go-mutesting mutation testing tool.