Wait for watch.Value subscription to finish
The new documentation for Cancel, along with the new Wait function,
resolves the issue I noticed last night about Cancel's guarantee of no
new handler executions not holding up under some interleavings.

Doing this as a function is a bit of an extra burden on consumers that
heavily use channels, but I figure this API is cleaner than the
Context-like Done() function I originally introduced.