Provide new WebRTC logic to existing client
This change completes the rewrite of the backend API, providing the last
unimplemented route. The existing client has been cut over to the new
backend, since with the removal of other logic from the existing
endpoint this was a fairly simple change that only affected the SDP
message format on the wire.