Initial changes for H.264 encoding
With this change, we can successfully encode the video stream as H.264
and play it back on Chrome. As expected, x264 does a far better job of
realtime encoding than libvpx, and the quality of the stream is
noticeably better.

The current problem is that Firefox refuses to accept the H.264 stream,
even though H.264 is the only encoding option provided in the server's
session description. It literally sends back an answer where the video
stream is inactive and the only rtpmap entry is for VP8.

I know that when I set up a new RTCPeerConnection, add a recvonly video
transceiver, and create an offer in a Firefox dev console, the generated
description includes rtpmap entries for H.264. So, my guess is that I
need to make the initial call from the client rather than the server.
That makes sense to me intuitively, at least. WebRTC requires that you
support *both* VP8 and H.264, so technically I'm the one in violation of
the spec by ignoring Firefox's request to use VP8. Whereas if Firefox is
the caller it has to give me the option to choose, and I'm allowed to
choose H.264.