Even more fixes and updates to the watch package
A few additional updates to the watch package, since it plays such a
pivotal role in the implementation.

First, the term "subscribe" (and variants) has consistently been
replaced with the term "watch." For one thing, this is more natural to
write as a method name. Also, I think it more clearly captures the idea
that you'll get an initial call with the current value of a Value after
starting your watch.

Second, the implementation has been updated to provide more predictable
behavior for watchers. Most notably, it is no longer possible for a
single watcher to see the same call to Set more than once, which may
simplify things for certain watchers (it is still possible to lose
updates). Also, a bug that may have resulted in lost updates when first
creating a watch has been fixed thanks to additional locking.

Documentation, variable names, etc. have been updated as necessary.