Remove unnecessary lock
I had this ridiculous belief over the weekend that I needed to hold the
value lock before taking the watchers lock to prevent deadlocks. It's
true that the two locks have to be consistently taken in the same order,
but it's perfectly valid to take the watchers lock without also taking
the value lock. What is wrong with me?