Use 100% for viewport height
100vh has some issues in mobile browsers whose address bars dynamically
show and hide themselves. 100% exposes some weird issue in Firefox for
Android where it keeps setting the page height shorter than it's
supposed to and displaying a white region (not even our background
color) at the bottom of the page, but this doesn't seem to reproduce
anywhere else so I guess I'll just live with it.