Update to enable multi-platform builds
This change helps enable slightly smoother multi-platform builds with
BuildKit, should that ever be desirable. Specifically we only build the
client assets using the builder node's architecture, since the output is
the same either way.

I've tested this with `podman build`, and at the very least it doesn't
fail on the `FROM --platform=` syntax.