Clean up properly on websocket send errors
Previously, if an error occurred while sending a message to a websocket
client, the handler would try to shut down. However, if the error didn't
also cause the client read loop to break, the handler would simply get
stuck forever until the client broke the connection.

The handlers are now rearranged so that we don't wait for background
goroutines to finish until after we've forcibly shut down the websocket
connection. So we still wait for proper cleanup but there's now a
stronger guarantee that it will work.