Update Dockerfile to use BuildKit caching
This changeset takes advantage of the `RUN --mount` Dockerfile syntax
provided by BuildKit to simplify the container image build process and
make it more efficient. In particular, I prefer cache mounts to the
"traditional" strategy of carefully layering COPY and RUN commands,
which is more sensitive to changes in file structure and busts the
entire "cache" on every file change.

The tricky part of BuildKit bind mounts is that you can't write to them
(writes to a rw mount are dropped when the command finishes). So all
build output has to be written outside of the mounted directory.