Pass GstBuffer directly to Go
This means we no longer need to pass the duration as a separate
argument, which I think is easier to read and understand. It also lets
us extract the GstBuffer data directly into a Go byte slice, which is
one fewer C allocation for human beings to think about. (This is NOT
intended as a performance or efficiency improvement; the only truly
expensive part of Hypcast is H.264 encoding.)