Minor updates to C style
In https://blog.golang.org/declaration-syntax, Rob Pike provides a
notable insight into C declarations:

  C took an unusual and clever approach to declaration syntax. Instead
  of describing the types with special syntax, one writes an expression
  involving the item being declared, and states what type that
  expression will have.

This makes it clear to me that moving the "*" next to the name of the
variable, rather than next to the name of the type, is the superior
style for C code (as opposed to Go code). From the C style guides I've
reviewed that express opinions on this matter, Google's (interestingly
enough) is the only one that does not explicitly prefer this style, but
instead leaves it up to the programmer. I see this as further evidence
that this is an appropriate change to make.