Pass GstSample directly to Go
This is like the extended version of the earlier GstBuffer change. It
further simplifies the interface from GStreamer into Go (no longer
separating buffer and size), and is more in line with the general
pattern of keeping as much of the logic as possible in Go source, even
when that means more C calls from Go.

As part of this the liveness of the GstSample (and by extension the
GstBuffer) is now much more tightly controlled.  But once again, this
change is NOT intended to be an "optimization" of any sort, and is
purely about my subjective ability to understand and maintain this code.

More importantly, as part of thinking carefully about GstSample
liveness, I fixed a bug where a null GstBuffer could result in a memory
leak due to not unref-ing the GstSample.