Support arbitrary appsink names
This should effectively complete the refactor of the gst package, by
allowing consumers to set up sink functions for arbitrary named appsink
elements.

While writing this change, I learned that the return value for
gst_bin_get_by_name is "transfer full," and that I was probably leaking
memory before by forgetting to unref the bin. That is now fixed.