Insulate main watch loop from goroutine exits
After fixing the invalid uses of T.Fatal in my watcher tests, I realized
that in general, calling runtime.Goexit from a watch handler would cause
the watch to stop operating properly. I don't think this is reasonable
behavior, particularly given the pre-existing documentation that the
watch handler simply runs "in a new goroutine."

Interestingly, the new implementation causes the Set benchmarks to run
*faster* for the cases of 10 and 100 watchers. My suspicion is that it
now hits the case where it updates the watch's pending value more often,
rather than the case where it provides a new pending value to an empty
channel, and perhaps the latter requires more interaction with the
runtime to mark the waiting goroutine for the main run loop as ready to
be scheduled. In other words, Set might be faster since it drops
intermediate values on the floor more often, while new values actually
are slower to be picked up, however we just don't have a benchmark for
that latter case. (Obviously the change is to improve correctness and I
have no *real* interest in the resulting benchmarks, I just thought this
was a fun observation.)