Simplify watch implementation with one lock
I've been reading up on mutexes lately, and particularly was curious why
Go doesn't implement recursive mutexes, because for some weird reason I
thought that they might have been potentially useful somewhere (I forget
where).

https://groups.google.com/g/golang-nuts/c/XqW1qcuZgKg/m/Ui3nQkeLV80J

https://groups.google.com/g/comp.programming.threads/c/tcrTKnfP8HI/m/me2K7_byNdgJ

https://blog.stephencleary.com/2013/04/recursive-re-entrant-locks.html

I learned something important from these 3 domain experts, that I hadn't
truly understood before: the purpose of a mutex is to *protect
invariants* across different fragments of your overall state. (Also,
that recursive mutexes are almost always a bad tool to use.)

Thinking about this package, I recognized that the key invariant is the
one protected by valueMu, which involves the relationship between the
value and the set of watches that need to receive that value. When
removing a watch, the necessary invariant can be maintained with an
atomic update to the watchers set alone, which I previously implemented
using the separate watchersMu.

In practice, though, 2 of the 3 operations that affect watches required
*both* a lock on valueMu to protect the invariants for Value as a whole,
and a lock on watchersMu to protect the invariants specific to Go's map
implementation. Given that the internal invariants of the map are part
of the overall invariants required for the Value implementation, it's
clear that the same task can be accomplished with a single mutex, with
significantly less complexity in the implementation.

In fact, where I might have naïvely thought that the more "independent"
locks would be better for "performance" or "efficiency" or whatever, I
suspect that the single lock is actually better on those fronts, since
it involves fewer potential interactions with the runtime. If nothing
else, my crappy Set benchmark *seems* to run, like, 20ns faster with the
single-lock implementation. (Keep in mind that the benchmark doesn't
involve any updates to the watchers set alongside those Set calls.)

Alongside this change I've also renamed and cleaned up some other parts
of the implementation in very minor ways.