Hack in vaapih264enc support
Unlike NVENC, Intel's H.264 hardware encoder supports the Constrained
Baseline profile.

To be honest, I'm not as impressed with the quality of the encoding as I
am with NVENC, but I think with these settings it's acceptable.

I also had to set some GStreamer and libva environment variables to use
my Intel hardware instead of the NVIDIA card that drives my displays, so
it's not quite as zero-configuration as the NVENC setup was for me, and
certainly not as zero-configuration as x264.