Enable client embedding via build tags
This makes it easier to build and run Hypcast on a system where the
container image is hard to use, e.g. because you want to use hardware
encoding elements that aren't easy to install on Alpine Linux. Running
"make" from a clean checkout should do everything necessary to produce a
working build as long as all the right dependencies are in place.