Prototype full VA-API video transcoding pipeline
This pipeline successfully moves the MPEG-2 decoding and video
deinterlacing steps of the pipeline over to VA-API, enabling the
possibility of end-to-end hardware acceleration for all video
processing. GStreamer is smart enough to keep everything in video memory
until it finally gets passed to the application.

With this in place, instead of swapping out just the encoding element, I
want an option to enable or disable VA-API acceleration as a whole.
Default to software, but use VA-API if available.