Enable client embedding via ZIP file
I've wanted to try this for a while. My real end goal is to have a HTTP
handler that extracts the gzip streams straight out of the archive to
serve them to gzip-capable HTTP clients, since I've always thought that
would be a neat idea and am personally curious about how it would work.