Implement minimal gzip serving in zipserve
When a browser indicates that it accepts gzip encoding, zipserve will
now serve the compressed content straight from the ZIP file just as
advertised.

The implementation is really crappy, e.g. we need to actually parse the
header correctly instead of just doing a string match, but this is
enough to show that the basic concept really does work.