"Cleanup" and more consistency for zipserve
This package is still a horrifying mess, but at least now I've named
some of the more magical constants and made some of the functionality
better (e.g. doing Content-Length properly on all requests).

I think I need to let this sit for a while and come back much later for
a fresh look on how to factor this logic properly.