Force software MPEG-2 decode in default pipeline
decodebin gave GStreamer the flexibility to select the VA-API MPEG-2
decoder element, which in my setup only works if I explicitly override
the driver and DRM device, which is why I made the VA-API pipeline a
non-default option in the first place. With this change the default
pipeline should behave the way it did originally before I installed the
VA-API plugins.