Remove unused "raw" appsink
I don't see a need to keep maintaining this until I actually find a use
case for it. The original idea was to parse MPEG in Go (alongside
GStreamer) to try to extract program information, however the balance
between how easy it is to do and how much I actually care about the
feature hasn't worked out in favor of keeping this, at least for the
moment.