First prototype of closed caption support
This implementation can display CEA-708 closed captions on a fairly
small number of channels. However, some channels don't seem to display
the captions at all, even though they should have them, and in some
cases (at least with the VA-API H.264 encoder) it causes massive issues
with audio and video getting out of sync, or streams not starting up at
all.

It took a ton of effort even to figure out how to get the ccoverlay
element working, with all the right queues and links in place. It's also
hard for me to tell where exactly the remaining issues are:
cc708overlay, vaapih264enc, etc. I'm not sure if I'm still doing
something wrong, or if this is just an unreliable feature set in
GStreamer. So, unfortunately, I'm not sure if it makes sense to keep
trying to go down this path.

It might be interesting to consider this strategy:
https://dvcs.w3.org/hg/text-tracks/raw-file/default/608toVTT/608toVTT.html.
The document talks about converting CEA-708 data streams to WebVTT,
which... is probably very hard. But I suspect also a lot more reliable.