Use cgo.Handle to reference gst.Pipeline in C
This new feature from Go 1.17 obviates the need for a custom
implementation of the "big old map" concept.

I might be able to take this even farther by directly making a handle to
a sinkDef and shoving that handle value into a gpointer. GLib itself has
GSIZE_TO_POINTER and GPOINTER_TO_SIZE macros that do this kind of
conversion, so I assume it's cool to play that kind of game even though
you have a "pointer" that will instantly explode if you dereference it.
In fact, when you think about it, if GStreamer actually tried to
dereference these user data pointers internally you would never be able
to pass a null pointer, which seems like an odd restriction. So as long
as you're very careful to do these conversions at the last possible
second I feel like it has to technically work.