Use type parameters for watch.Value
This is the first real code I've written with type parameters in Go,
originally with the pre-1.18 master branch of Go but now using the
official 1.18 release series. (In fact, this specific code uncovered
multiple issues with type parameter support shortly after it landed in
the Go master branch; a special thanks to the Go team for their quick
and friendly responses to the various issues that I and others filed!)

I had to parameterize the underlying watch type to ensure that the
pending channel and callback fields have the correct types. Since the
Cancel and Wait methods don't need to care about the type parameter,
I've turned watch.Watch into an unparameterized interface type so that
clients of the package don't have to instantiate a type for every watch
they store in a struct field (or a variable whose type is declared
explicitly).

I've also taken the liberty of cleaning up the code in a variety of
other ways, particularly the unit tests. I've removed the completely
ridiculous and useless "benchmark" that never should have been written
in the first place, since the results probably just reflect lock
contention.