Start restructuring zipserve to use fs.FS
I came up with what I think is a better idea for zipserve: Back it with
a fs.FS instead of just a ZIP archive, and look for .gz files in order
to serve pre-compressed content. This way it can support on-disk content
in addition to ZIP archives, so that the behavior of asset serving is
more consistent. On the ZIP archive side, the plan is to implement my
own fs.FS that exposes .gz files alongside the regular entries in a ZIP
archive, assuming there isn't already a .gz entry present in the
archive.

Right now the behavior is kind of similar to before, except that the
handler happily tries to serve directories as if they're normal files.