Clean up the rpc handler a little
First, rearrange some of the code to a form that I think is a bit easier
to understand. I especially think the early exit on an empty body is a
lot cleaner than duplicating the `body != nil` check.

Second, replace the custom zero function with `*new(T)`. This was sort
of an experiment on my part to imagine a syntax that's a little nicer to
read than `*new(T)`, more like a hypothetical `zero(T)`. But that was
never really clear, and I might as well just do what everyone else does.