Introduce emulation-free multi-arch Docker builds
This change reworks the Dockerfile to support builds for the
linux/amd64, linux/arm64, and linux/arm/v7 platforms, without requiring
any binfmt_misc hacking or emulation with QEMU. With this change it
should be possible to build a multi-arch Hypcast image on any Linux
system that can run Docker, such as (hint hint) a GitHub Actions worker.

All of this is accomplished using Alpine Linux and an LLVM-based
toolchain (clang and lld). The apk package manager provides built-in
support for operating on chroots across architectures, and can create
extremely minimal sysroot directories containing just basic libraries
and headers with relative ease. LLVM provides cross-compilation support
for many architectures in a single set of host binaries, much like Go
itself (and unlike GNU toolchains that are built individually per host
and target combination).

Part of what I like about this approach is that I'm still entirely using
Alpine Linux and its package ecosystem, rather than manually downloading
a toolchain or image that I might not be as familiar with, or that is
much larger than what I actually need. I considered a switch back to
Debian, using the goreleaser-cross image instead of my current bases,
and even some of the approaches for building cgo applications with Zig.
But out of all of those options, this is what I think I'd be most
comfortable debugging and maintaining.

Another nice element is that thanks to how BuildKit, Go, apk, and LLVM
are all great at handling a wide variety of architectures with, the same
overall process is used for both native and cross builds with no
special-casing for the current host architecture. This gives me much
more confidence that the setup I've tested on a Mac with Apple silicon
should readily transfer to a GitHub Actions worker or any other
BuildKit-enabled host, and that if any random person attempts a normal
`docker build --platform` they should have no trouble with the process.

The only big issue I had was when lld linked a position-independent
executable for x86_64 by default while Go was still building the runtime
and other components as non-PIE. Unsurprisingly, the Go runtime really
does not like that situation and crashes before it can even initialize
the heap. I opted to enable PIE in Go rather than disable it in LLVM,
since it's generally how all Alpine packages are built (which might well
be why Alpine's LLVM toolchain enables it by default).