Go distroless; don't run apk build scripts
For whatever reason, it didn't occur to me at all that apk would run
scripts using the shell inside of the chroot, which handle things like
setting up all of the symlinks for Busybox tools.

I've gone ahead and disabled build scripts, and since we're not going to
have anything resembling a standard Alpine environment anyways I've
scoped the packages for the target chroot down to the bare minimum. I
don't even list musl as a dependency, I'm relying on standard package
dependencies to make sure it's installed. Technically we do still have
Busybox in the final container, and it even comes with a /bin/sh
symlink, so it's not 100% like a typical distroless image. But it is
awkward to shell into and doesn't include a package manager, so it's
definitely a special-purpose kind of thing.