Roll back to Alpine 3.15 for mpeg2dec
The default build of gst-plugins-ugly in Alpine 3.16 disables the
mpeg2dec plugin as "perhaps less useful," which is unfortunate for
Hypcast given that MPEG 2 decoding is a very significant portion of what
it needs to do. Alpine 3.15 will continue to receive security fixes
until 1 November 2023, but I don't know what the solution is going to be
after that.