Add initial nhooyr.io/websocket dependency
I've recently heard about the Gorilla Toolkit being archived. While
gorilla/websocket has worked quite well, and should continue to work
well even after the archival of the project, the event did convince me
to look into what other options are available, and there are some things
I like about this alternative package (including the built-in function
to drain readers, and the built-in function to decode JSON messages with
a buffer pool; it is the typical naïve buffer pool implementation that
holds on to more memory than necessary after handling a single large
message, but this is mitigated somewhat by the new library having a
default 32 kiB read size limit, which itself seems smart to me).

If nothing else I think it's worth a trial just to keep my knowledge of
the ecosystem fresh and get a sense of different ways of doing things.