Use new websocket package for tuner status
This is the easy one, since it's write-only.

I decided *not* to use CloseRead in the end, and to keep draining the
readers myself. This gives a little more control over how to shut down,
and keeps the previous behavior of ignoring messages instead of closing
with a protocol violation (though the latter is probably the better
move).