Use new websocket package for WebRTC signaling
This is the final change that removes all remaining gorilla/websocket
usage within Hypcast itself (the new websocket package, ironically
enough, uses it for some of its tests, which is why it's still an
indirect dependency in go.mod).

The net diff is more than I think I was hoping for, but some of that is
because we now treat the client going away as a non-error shutdown
rather than an error shutdown. We also have more of an ability to use
contexts for cancellation, which might really clean up the shutdown code
when combined with WithCancelCause in Go 1.20.