Move LLVM cross options to Meson binaries section
There are a couple things on the Meson issue tracker that imply this is
the way to go, since it can avoid issues with some aspects of Meson's
feature detection targeting the build system instead of the host system,
or something along those lines. In any case it makes the config smaller,
which seems nice I guess?