Make the clickable elements look more clickable
I never really thought about this before, but I don't think it's
immediately intuitive that the power button is actually a useful control
and not just a big status icon. The channel list is a bit more obvious
(since how the heck else would you tune to a channel), but I still think
it looks nicer with some hover styles and a pointer cursor.

Also, I never actually documented where I got my palette shades from,
which is kind of an issue when there's a use case to add more.
Thankfully I actually remembered the site and even managed to get a link
that pre-populates our two base colors.