Initial hacky integration of gzipped.FileServer
With this, I might have a path to do this weird thing I've wanted to try
for a long time: serve the DEFLATE stream *directly* out of the embedded
client .zip file.

This is a very weird idea, and not something I'd particularly recommend
in production-grade code, but since this is a personal project I figure
I can have a bit of leeway to try ridiculous things like this.

In unrelated news, it turns out that http.FileSystem requires all files
to implement io.Seeker, and that I've been breaking this requirement for
a long time with the .zip embedding approach, since those files
obviously don't support seeking.