Experiment with watches that don't pre-spawn goroutines
What does it take to implement watches without long-running goroutines?
This is a first pass based on some work I did in another project, which
itself is based on patterns from Bryan Mills' Gophercon 2018 concurrency
patterns talk, particularly the idea of a 1-buffered channel acting like
a mutex over some data.

This pattern has its benefits, but it does feel awkward to be so careful
about sending the state back into the channel at the right times. In my
other project this work had the benefit of eliminating the need for a
Close()-type method, which was a welcome API simplification, but that
doesn't really apply here.

This might be more interesting if I loosen the requirement around Watch
calling the handler "with the value stored in v upon creation of the
watch," and make it so the first value can come from a later call to
Set. The idea, of course, is not to store copies of the value in each
watch. I'm still thinking through what this might entail.